+++ docs/styles/layout.md
@@ -1,0 +1,13 @@
+# Layout
+
+## Positioning
+
+### Relative Positioning
+
+### Absolute Positioning
+
+```python
+--8<-- "absolute_positioning.py:example"
+```
+
+![Absolute Positioning](../examples/absolute-positioning.svg)
